Redesign of the extension modal and underlying import functions
The new modal contains a separated navigation section with search functionality to browse the various object types on the left. On the right of the modal a single object's details and preview are displayed.
Importing of app objects is rewritten to only process properties instead of the occasional object layout data.
Updating of app objects is provided for those objects that already exist in the current app. App objects are matched on object titles.

Importing bookmarks is removed from the extension's functionality as there is no reliable way found to support this.